[Design] 퀘스트 탭에 사용되는 퀘스트 카드를 기본 퀘스트 카드로 수정,
